Refactoring of file related utilities.
